+++ crates/slosh3d/examples/heightfield3.rs
@@ -1,7 +1,8 @@
 use slosh_testbed3d::{RapierData, slosh};
 
-use nalgebra::{DMatrix, point, vector};
+use glam::vec3;
 use rapier3d::geometry::{HeightField, TriMeshFlags};
+use rapier3d::parry::utils::Array2;
 use rapier3d::prelude::{ColliderBuilder, RigidBodyBuilder};
 use slang_hal::backend::WebGpu;
 use slosh::{
@@ -24,11 +25,11 @@ pub fn heightfield_demo(backend: &WebGpu, app_state: &mut AppState) -> PhysicsCo
     for i in 0..nxz {
         for j in 0..100 {
             for k in 0..nxz {
-                let position = point![
+                let position = vec3(
                     i as f32 + 0.5 - nxz as f32 / 2.0,
                     j as f32 + 0.5 + 14.0,
-                    k as f32 + 0.5 - nxz as f32 / 2.0
-                ] * cell_width
+                    k as f32 + 0.5 - nxz as f32 / 2.0,
+                ) * cell_width
                     / 2.0;
                 let density = 2700.0;
                 let radius = cell_width / 4.0;
@@ -45,14 +46,14 @@ pub fn heightfield_demo(backend: &WebGpu, app_state: &mut AppState) -> PhysicsCo
     };
 
     let params = SimulationParams {
-        gravity: vector![0.0, -9.81, 0.0] * app_state.gravity_factor,
+        gravity: vec3(0.0, -9.81, 0.0) * app_state.gravity_factor,
         dt: 1.0 / 60.0,
     };
 
-    let heights = DMatrix::from_fn(200, 200, |i, j| {
+    let heights = Array2::from_fn(200, 200, |i, j| {
         (i as f32 / 10.0).sin() * (j as f32 / 10.0).cos()
     });
-    let heightfield = HeightField::new(heights, vector![100.0, 5.0, 100.0]);
+    let heightfield = HeightField::new(heights, vec3(100.0, 5.0, 100.0));
     let (vtx, idx) = heightfield.to_trimesh();
     let rb = RigidBodyBuilder::fixed();
     let rb_handle = rapier_data.bodies.insert(rb);